Adds the External Identifier driver.
The External Identifier driver is unique in that all views detect if it is enabled for a given protocol. If it is, all external records for that datasource are pulled and rendered on each downstream view. It should be noted that the URL for the External Identifier driver is not important as no downstream record creation is done (only the eHB is being used to store IDs). The URL, however, must match the External System defined for this driver in the eHB.

Fixes #11.