Removing jumpahead function.
The jumpahead function was removed in Py3 because it was deemed ineffective.

See: http://bugs.python.org/issue9816#msg115990

'The function/method was removed in 3.x because it was an API defect.  The jumpahead concept as originally intended (move ahead n-steps) was something that could really only work with a generator like Wichmann-Hill.  Newer and more advanced generators aren't usually amenable to direct computation of a state that is n-steps forward.'